+++ CHANGELOG.md
@@ -16,10 +16,10 @@
 - jslint - try to improve parser to be able to parse jquery.js without stopping.
 - merge function.html and help.html into README.md
 - node - after node-v14 is deprecated, remove shell-code `export "NODE_OPTIONS=--unhandled-rejections=strict"`.
-- vim - add vim plugin.
 
 # v2021.9.1-beta
 - jslint - add bigint support.
+- vim - add vim plugin.
 
 # v2021.8.20
 - warning - disable un-ergonomic warnings restricting directive-global (missing_browser and unexpected_directive_a).